Move TimeDerivStrahlkorper to StrahlkorperFunctions
