+++ src/NumericalAlgorithms/SphericalHarmonics/StrahlkorperFunctions.cpp
@@ -3,11 +3,16 @@
 
 #include "NumericalAlgorithms/SphericalHarmonics/StrahlkorperFunctions.hpp"
 
+#include <deque>
+#include <utility>
+
 #include "DataStructures/DataVector.hpp"
 #include "DataStructures/Tensor/Tensor.hpp"
 #include "DataStructures/Tensor/TypeAliases.hpp"
+#include "NumericalAlgorithms/FiniteDifference/NonUniform1D.hpp"
 #include "NumericalAlgorithms/Interpolation/LinearLeastSquares.hpp"
 #include "NumericalAlgorithms/SphericalHarmonics/Strahlkorper.hpp"
+#include "Utilities/ErrorHandling/Error.hpp"
 #include "Utilities/GenerateInstantiations.hpp"
 #include "Utilities/Gsl.hpp"
 #include "Utilities/SetNumberOfGridPoints.hpp"
@@ -81,9 +86,8 @@ StrahlkorperTags::aliases::Jacobian<Fr> jacobian(
 }
 
 template <typename Fr>
-void jacobian(
-    const gsl::not_null<StrahlkorperTags::aliases::Jacobian<Fr>*> jac,
-    const tnsr::i<DataVector, 2, ::Frame::Spherical<Fr>>& theta_phi) {
+void jacobian(const gsl::not_null<StrahlkorperTags::aliases::Jacobian<Fr>*> jac,
+              const tnsr::i<DataVector, 2, ::Frame::Spherical<Fr>>& theta_phi) {
   const auto& theta = get<0>(theta_phi);
   const auto& phi = get<1>(theta_phi);
   set_number_of_grid_points(jac, theta);
@@ -347,7 +351,7 @@ Scalar<DataVector> laplacian_of_scalar(
     const Scalar<DataVector>& scalar, const Strahlkorper<Fr>& strahlkorper,
     const tnsr::i<DataVector, 2, ::Frame::Spherical<Fr>>& theta_phi) {
   Scalar<DataVector> result{DataVector{get(scalar).size()}};
-  laplacian_of_scalar(make_not_null(&result),scalar,strahlkorper,theta_phi);
+  laplacian_of_scalar(make_not_null(&result), scalar, strahlkorper, theta_phi);
   return result;
 }
 
@@ -431,6 +435,65 @@ std::vector<std::array<double, 4>> fit_ylm_coeffs(
     coefficients.push_back(std::move(same_coeff_of_each_strahlkorper));
   }
   return predictor.fit_coefficients(times, coefficients);
+}
+
+namespace {
+template <size_t NumTimes>
+static DataVector compute_coefs(
+    const std::deque<double>& times,
+    const std::deque<const DataVector*>& coefficients) {
+  const auto weights = fd::non_uniform_1d_weights<NumTimes>(times);
+
+  DataVector new_coefficients{coefficients.front()->size(), 0.0};
+
+  for (size_t i = 0; i < NumTimes; i++) {
+    new_coefficients += *coefficients[i] * gsl::at(gsl::at(weights, 1), i);
+  }
+
+  return new_coefficients;
+}
+}  // namespace
+
+template <typename Frame>
+void time_deriv_of_strahlkorper(
+    gsl::not_null<Strahlkorper<Frame>*> time_deriv,
+    const std::deque<std::pair<double, ::Strahlkorper<Frame>>>&
+        previous_strahlkorpers) {
+  std::deque<double> times{};
+  std::deque<const DataVector*> coefficients{};
+
+  for (const auto& [time, strahlkorper] : previous_strahlkorpers) {
+    // This only happens toward the beginning because the first time is NaN and
+    // if that happens we can't actually take a derivative
+    if (UNLIKELY(std::isnan(time))) {
+      time_deriv->coefficients() =
+          DataVector{strahlkorper.coefficients().size(), 0.0};
+      return;
+    }
+
+    times.emplace_back(time);
+    coefficients.emplace_back(&strahlkorper.coefficients());
+  }
+
+  DataVector new_coefficients{};
+
+  // Switch needed to convert times size into template parameter
+  switch (previous_strahlkorpers.size()) {
+    case 2:
+      new_coefficients = compute_coefs<2>(times, coefficients);
+      break;
+    case 3:
+      new_coefficients = compute_coefs<3>(times, coefficients);
+      break;
+    case 4:
+      new_coefficients = compute_coefs<4>(times, coefficients);
+      break;
+    default:
+      ERROR("Unsupported size for number of previous Strahlkorpers "
+            << previous_strahlkorpers.size());
+  }
+
+  time_deriv->coefficients() = std::move(new_coefficients);
 }
 }  // namespace StrahlkorperFunctions
 
@@ -556,7 +619,10 @@ std::vector<std::array<double, 4>> fit_ylm_coeffs(
   template std::vector<std::array<double, 4>>                                  \
   StrahlkorperFunctions::fit_ylm_coeffs(                                       \
       const DataVector& times,                                                 \
-      const std::vector<Strahlkorper<FRAME(data)>>& strahlkorpers);
+      const std::vector<Strahlkorper<FRAME(data)>>& strahlkorpers);            \
+  template void StrahlkorperFunctions::time_deriv_of_strahlkorper(             \
+      const gsl::not_null<Strahlkorper<FRAME(data)>*>,                         \
+      const std::deque<std::pair<double, ::Strahlkorper<FRAME(data)>>>&);
 
 GENERATE_INSTANTIATIONS(INSTANTIATE,
                         (Frame::Distorted, Frame::Grid, Frame::Inertial))
